feat(start): Move start and start-available cmd to mip/base.py
