+++ cg/cli/workflow/mip/base.py
@@ -5,6 +5,7 @@ from typing import List
 import click
 
 from cg.apps.environ import environ_email
+from cg.cli.workflow.commands import ensure_flowcells_ondisk, link, resolve_compression
 from cg.cli.workflow.mip.options import (
     ARGUMENT_CASE_ID,
     EMAIL_OPTION,
@@ -15,7 +16,8 @@ from cg.cli.workflow.mip.options import (
     PRIORITY_OPTION,
     START_WITH_PROGRAM,
 )
-from cg.exc import CgError
+from cg.constants import EXIT_FAIL, EXIT_SUCCESS
+from cg.exc import CgError, DecompressionNeededError, FlowcellsNeededError
 from cg.meta.workflow.mip import MipAnalysisAPI
 
 LOG = logging.getLogger(__name__)
@@ -113,4 +115,73 @@ def run(
         LOG.info("%s run started!", analysis_api.pipeline)
     except CgError as e:
         LOG.error(e.message)
+        raise click.Abort
+
+
+@click.command()
+@ARGUMENT_CASE_ID
+@EMAIL_OPTION
+@OPTION_DRY
+@OPTION_MIP_DRY_RUN
+@OPTION_PANEL_BED
+@OPTION_SKIP_EVALUATION
+@PRIORITY_OPTION
+@START_WITH_PROGRAM
+@click.pass_context
+def start(
+    context: click.Context,
+    case_id: str,
+    dry_run: bool,
+    email: str,
+    mip_dry_run: bool,
+    panel_bed: str,
+    priority: str,
+    skip_evaluation: bool,
+    start_with: str,
+):
+    """Start full MIP analysis workflow for a case"""
+
+    analysis_api: MipAnalysisAPI = context.obj["analysis_api"]
+
+    analysis_api.verify_case_id_in_statusdb(case_id=case_id)
+    LOG.info("Starting full MIP analysis workflow for case %s", case_id)
+    try:
+        context.invoke(ensure_flowcells_ondisk, case_id=case_id)
+        context.invoke(resolve_compression, case_id=case_id, dry_run=dry_run)
+        context.invoke(link, case_id=case_id)
+        context.invoke(panel, case_id=case_id, dry_run=dry_run)
+        context.invoke(config_case, case_id=case_id, panel_bed=panel_bed, dry_run=dry_run)
+        context.invoke(
+            run,
+            case_id=case_id,
+            priority=priority,
+            email=email,
+            start_with=start_with,
+            dry_run=dry_run,
+            mip_dry_run=mip_dry_run,
+            skip_evaluation=skip_evaluation,
+        )
+    except (FlowcellsNeededError, DecompressionNeededError) as e:
+        LOG.error(e.message)
+
+
+@click.command()
+@OPTION_DRY
+@click.pass_context
+def start_available(context: click.Context, dry_run: bool = False):
+    """Start full MIP analysis workflow for all cases ready for analysis"""
+
+    analysis_api: MipAnalysisAPI = context.obj["analysis_api"]
+
+    exit_code: int = EXIT_SUCCESS
+    for case_obj in analysis_api.get_cases_to_analyze():
+        try:
+            context.invoke(start, case_id=case_obj.internal_id, dry_run=dry_run)
+        except CgError as error:
+            LOG.error(error.message)
+            exit_code = EXIT_FAIL
+        except Exception as e:
+            LOG.error(f"Unspecified error occurred: %s", e)
+            exit_code = EXIT_FAIL
+    if exit_code:
         raise click.Abort
